MAINTAINERS/OWNERS: Transfer from venture@ to brandonkim@
Patrick Venture has not been actively working on OpenBMC and seeing
as how this repository is only pulled in by phosphor-ipmi-flash, we
agreed that I (as a maintainer of that repository) would be a good
replacement.

Patrick will resign his maintainership while I will take over.
The plan is to ask and see if other phosphor-ipmi-flash maintainers
would also like to join the maintainer / owner / reviewer role.

Signed-off-by: Brandon Kim <brandonkim@google.com>
Change-Id: I8c676963d617f7fb1789bc485bc24d0f41cf0b89